requirements.txt
typer
pyaudio
openai-whisper
numpy
fastapi
python-multipart
pyyaml
pytest
pytest-cov